requirements.txt
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
black==23.11.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
coverage==7.3.2
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
hypothesis==6.92.0
identify==2.5.33
idna==3.6
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.18.1
jaraco.classes==3.3.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mypy==1.6.1
mypy-extensions==1.0.0
mypy-protobuf==3.3.0
nest-asyncio==1.5.8
nh3==0.2.15
nodeenv==1.8.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parameterized==0.9.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pip==23.3.1
pkginfo==1.9.6
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
prompt-toolkit==3.0.42
protobuf==3.20.3
psutil==5.9.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
readme-renderer==42.0
requests==2.31.0
requests-mock==1.11.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
semver==2.13.0
setuptools==68.0.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
stack-data==0.6.2
testfixtures==7.2.2
tornado==6.3.3
traitlets==5.14.0
twine==4.0.2
types-click==7.1.8
types-protobuf==4.24.0.4
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-requests==2.31.0.10
types-setuptools==69.0.0.0
types-toml==0.10.8.7
types-typed-ast==1.5.8.7
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
wheel==0.41.2
zipp==3.17.0
